// frontend/src/data/reference.generated.ts
export interface Trigram {
  id: string; name: string; hanzi: string; lines: number[];
  image: string; action: string; family: string;
  element: string; direction: string; classic: string;
}
export interface Hexagram {
  number: number; lower: string; upper: string;
  name: string; title: string; essence: string;
}

export const TRIGRAMS: Record<string, Trigram> = {
  "qian": {
    "id": "qian",
    "name": "Цянь",
    "hanzi": "乾",
    "lines": [
      1,
      1,
      1
    ],
    "image": "Небо",
    "action": "Творчество",
    "family": "Отец",
    "element": "Металл",
    "direction": "Северо-запад",
    "classic": "[TODO Таня]"
  },
  "kun": {
    "id": "kun",
    "name": "Кунь",
    "hanzi": "坤",
    "lines": [
      0,
      0,
      0
    ],
    "image": "Земля",
    "action": "Исполнение",
    "family": "Мать",
    "element": "Почва",
    "direction": "Юго-запад",
    "classic": "[TODO Таня]"
  },
  "zhen": {
    "id": "zhen",
    "name": "Чжэнь",
    "hanzi": "震",
    "lines": [
      1,
      0,
      0
    ],
    "image": "Гром",
    "action": "Возбуждение",
    "family": "Старший сын",
    "element": "Дерево",
    "direction": "Восток",
    "classic": "[TODO Таня]"
  },
  "xun": {
    "id": "xun",
    "name": "Сюнь",
    "hanzi": "巽",
    "lines": [
      0,
      1,
      1
    ],
    "image": "Ветер",
    "action": "Проникновение",
    "family": "Старшая дочь",
    "element": "Дерево",
    "direction": "Юго-восток",
    "classic": "[TODO Таня]"
  },
  "kan": {
    "id": "kan",
    "name": "Кань",
    "hanzi": "坎",
    "lines": [
      0,
      1,
      0
    ],
    "image": "Вода",
    "action": "Погружение",
    "family": "Средний сын",
    "element": "Вода",
    "direction": "Север",
    "classic": "[TODO Таня]"
  },
  "li": {
    "id": "li",
    "name": "Ли",
    "hanzi": "離",
    "lines": [
      1,
      0,
      1
    ],
    "image": "Огонь",
    "action": "Сцепление",
    "family": "Средняя дочь",
    "element": "Огонь",
    "direction": "Юг",
    "classic": "[TODO Таня]"
  },
  "gen": {
    "id": "gen",
    "name": "Гэнь",
    "hanzi": "艮",
    "lines": [
      0,
      0,
      1
    ],
    "image": "Гора",
    "action": "Незыблемость",
    "family": "Младший сын",
    "element": "Почва",
    "direction": "Северо-восток",
    "classic": "[TODO Таня]"
  },
  "dui": {
    "id": "dui",
    "name": "Дуй",
    "hanzi": "兌",
    "lines": [
      1,
      1,
      0
    ],
    "image": "Водоём",
    "action": "Разрешение",
    "family": "Младшая дочь",
    "element": "Металл",
    "direction": "Запад",
    "classic": "[TODO Таня]"
  }
};